🍗:: [#242] alert 분리
