+++ Projects/Modules/DSKit/Demo/Sources/DemoViewController/Alert/DSCaveatAlertViewController.swift
@@ -8,25 +8,7 @@ import Then
 
 import DSKit
 
-public class DSAlertViewController: UIViewController {
-    
-    let alertView1 = MGAlertOnlyTitleView(title: "링크가 복사되었어요").then {
-        $0.titleLabel?.font = UIFont.Pretendard.labelMedium
-        $0.titleLabel?.textColor = .white
-        $0.backgroundColor = DSKitAsset.Colors.gray800.color
-    }
-    
-    let alertView2 = MGAlertBoxView(title: "링크가 복사되었어요", subtitle: nil, icon: .error).then {
-        $0.titleLabel?.font = UIFont.Pretendard.labelMedium
-        $0.titleLabel?.textColor = .black
-        $0.backgroundColor = .gray
-    }
-    
-    let alertView3 = MGAlertBarView(title: "링크가 복사되었어요", subtitle: nil, icon: .done).then {
-        $0.titleLabel?.font = UIFont.Pretendard.labelMedium
-        $0.titleLabel?.textColor = .black
-        $0.backgroundColor = .gray
-    }
+public class DSCaveatAlertViewController: UIViewController {
     
     let caveatAlertButton = MGButton(
         titleText: "caveatAlert",
@@ -37,11 +19,9 @@ public class DSAlertViewController: UIViewController {
         super.viewDidLoad()
         
         view.backgroundColor = .white
-        view.addSubview(caveatAlertButton)
-        alertView1.present(on: self.view)
-        alertView2.present(on: self.view)
-        alertView3.present(on: self.view)
         
+        view.addSubview(caveatAlertButton)
+
         caveatAlertButton.snp.makeConstraints {
             $0.centerX.centerY.equalToSuperview()
             $0.height.width.equalTo(200.0)
